requirements.txt
Django==1.8
django-cors-headers==2.0.2
djangorestframework==3.5.3
heroku==0.1.4
python-dateutil==1.5
requests==2.13.0
